price changed to positive integer field

+++ orders/models.py
@@ -24,7 +24,7 @@ class OrderItem(models.Model):
     order = models.ForeignKey(Order, on_delete=models.CASCADE, related_name='items')
     product = models.ForeignKey(Product, on_delete=models.CASCADE)
     quantity = models.PositiveIntegerField(default=1)
-    price = models.DecimalField(decimal_places=2, max_digits=5)
+    price = models.PositiveIntegerField()
 
     def __str__(self):
         return f'OrderItem: product: {self.product}, quantity: {self.quantity}, price: {self.price}'
